ai-service/requirements.txt
fastapi
uvicorn
neo4j
python-dotenv
langchain
langchain-community
langchain-core
langchain-neo4j
langchain-openai
langchain-google-genai
langchain-cohere
google-generativeai
openai
pydantic
requests
httpx
python-multipart
bcrypt
python-jose[cryptography]
email-validator
python-docx
PyPDF2
json-repair
bleach
# libsql-experimental (Requires Rust toolchain to compile, fallback to sqlite3 is implemented)
# libsql-experimental
cohere
typing-extensions
